tests/wpt/tests/html/browsers/history/the-history-interface/history_replacestate_too_many_calls_ordering.optional.html

<!DOCTYPE html>
<html>
<head>
<title>replaceState too many calls ordering</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
const maxCalls = 500;

async function createIframe(t) {
  const iframe = document.createElement("iframe");
  iframe.src = "/common/blank.html";
  t.add_cleanup(() => iframe.remove());

  const iframeLoaded = new Promise(resolve => {
    iframe.addEventListener("load", resolve, { once: true });
  });
  document.body.append(iframe);
  await iframeLoaded;
  return iframe;
}

function reachReplaceStateRateLimit(iframeWindow) {
  for (let i = 0; i < maxCalls; ++i) {
    const previousURL = iframeWindow.location.href;
    iframeWindow.history.replaceState(null, "", `#rate-limit-${i}`);
    if (iframeWindow.location.href === previousURL) {
      return;
    }
  }

  assert_unreached("No rate limit reached.");
}

promise_test(async t => {
  const iframe = await createIframe(t);
  const iframeWindow = iframe.contentWindow;
  reachReplaceStateRateLimit(iframeWindow);

  assert_throws_dom("DataCloneError", iframeWindow.DOMException, () => {
    iframeWindow.history.replaceState(() => {}, "", "#after-rate-limit");
  });
}, "replaceState serializes state before checking the rate limit");

promise_test(async t => {
  const iframe = await createIframe(t);
  const iframeWindow = iframe.contentWindow;
  reachReplaceStateRateLimit(iframeWindow);

  assert_throws_dom("SecurityError", iframeWindow.DOMException, () => {
    iframeWindow.history.replaceState(null, "", "http://[");
  });
}, "replaceState parses the URL before checking the rate limit");

promise_test(async t => {
  const iframe = await createIframe(t);
  const iframeWindow = iframe.contentWindow;
  reachReplaceStateRateLimit(iframeWindow);

  assert_throws_dom("SecurityError", iframeWindow.DOMException, () => {
    iframeWindow.history.replaceState(null, "", "https://example.com/");
  });
}, "replaceState checks whether the URL can be rewritten before checking the rate limit");

promise_test(async t => {
  const iframe = await createIframe(t);
  const iframeWindow = iframe.contentWindow;
  const iframeHistory = iframeWindow.history;
  const iframeDOMException = iframeWindow.DOMException;
  iframe.remove();

  for (let i = 0; i < maxCalls; ++i) {
    assert_throws_dom("SecurityError", iframeDOMException, () => {
      iframeHistory.replaceState(null, "", `#inactive-${i}`);
    });
  }
}, "replaceState checks fully active before checking the rate limit");
</script>
</body>
</html>
